Fix bug 605
From Juno release, when attaching/detaching volumes, if there is
an error raised (let's call it exception A), the work flow goes
into rollback process. If there is another exception raised in
rollback process (let's call it exception B), it will be exception
B printed in log file, and exception A will be totally lost.

Change-Id: Id5eb5f992229dc956659017066a2a071f3e7a5ba