Change the way to access block_device_mapping properties
The way to access properties of block_device_mapping object is
changed, so change the code to adapt to it. Block_device_mapping
used to be an object, so we access its properties by bdm.prop_name.
It becomes a dict now, so we have to access its properties by
bdm['prop_name']. The old way leads to exceptions.

Change-Id: I7a007c778a00e15d8c49f669523d8833b605b66b