Update restrictions of vm_state when attach/detach_volumes
If boot an instance with (not from) a volume, the state of the
instance will be BUILDING when attaching the volume. So we have
to allow attaching in that state. And detaching a volume from
a paused instance is not permitted.

Change-Id: Ic08aa699e6cc83bbd4df1d89ef379f20844ee574
(cherry picked from commit 5712d92ade6bdd3c3d4f7d2b03f4e771ebe2da7a)