Performance enhance to remove couple operation.
Reduce one smcli call for user direct to improve performance.

The basic idea behind this change is during spawn, there is no
reason to keep user direct temp state for NIC, we can wait for
all NIC info to be ready then add into a whole bunch.

so, all NIC info will be keep in db first, until neutron agent
found it's ok to add that NIC, it will add to the user direct
and nova side will wait for that complete then proceed.

This patch is work with neutron-zvm-agent's patch:
https://review.openstack.org/306243

Change-Id: I4ef246544c9c19a4941e045a6b1445e1041c256c