Support genconfig file
Currently the configuration are in driver.py, but not automatically
generated during build phase.

This patch enables 'tox -egenconfig' to build a sample conf file
and put into etc/nova-zvm.conf.sample, so if we need, we can paste
those info into nova.conf.sample when we do further build.

Follow up patches need to adjust the words in driver.py so
the auto generated conf.sample will be more helpful to user.

Change-Id: Ic35ecc8331c03bfe22e3077a8387ddfc1c8fb3d0