Use https with CA conf
CA file is a new config introduced to be used by compute service.
when compute service want to communicate with xcat, it add the
CA file into https communcation so it can verify the xcat is the
one compute is really the one it want to communicate with.

see following link for more info.
https://bugs.launchpad.net/ossn/+bug/1188189

Change-Id: I465a40ebb7486d03d16627948661e2d3a7dcf2f6