Modify the default shell used in punched files
Currently in some of our punched files we use /bin/sh to execute
shell script, on rhel and sles, the default shell is bash, and
/bin/sh is a symbolic link to /bin/bash,however, in Debian is no
longer bash, but dash.  /bin/sh is a symbolic link to /bin/dash
on newer Debian systems. It will cause some script fail to exec
so change all of them to use bash.

Change-Id: I8263912a79eb7be6d4c0e02676ddfb81978ec93f