Update root_disk_units with CYL or BLK units
Now the root_disk_units just has disk size in it. It doesn't have
units. So if customer specify a dismatch disk type, the driver
will not detect it until the create disk prompt the space is not
enough. This patch add units to root_disk_units, so it can give
customer clearly advise if error happens.

Change-Id: I1a31aa4bf77b774092fd04cc7174599252757538